+++ core/directory/Query.php
@@ -1,8 +1,10 @@
 <?php
 
+// phpcs:disable WordPress.DB.SlowDBQuery.slow_db_query_meta_query
+
 namespace Dev4Press\Plugin\GDMED\Directory;
 
-use Dev4Press\v51\Core\Quick\Sanitize;
+use Dev4Press\v54\Core\Quick\Sanitize;
 use stdClass;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -10,31 +12,29 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 class Query {
-	private $_r = array();
+	private array $_r = array();
 
-	private $_f = array(
+	private array $_f = array(
 		'orderby' => 'name',
 		'order'   => 'ASC',
 		'search'  => '',
 		'role'    => '',
 	);
 
-	/** @var null|\Dev4Press\Plugin\GDMED\Directory\MemberQuery */
-	private $_query = null;
+	private ?MemberQuery $_query = null;
 
-	/** @var null|\stdClass */
-	private $_pager = null;
+	private ?stdClass $_pager = null;
 
 	/** @var null|\Dev4Press\Plugin\GDMED\Directory\Member|\WP_User */
 	private $_member = null;
 
-	public $members = array();
-	public $members_ids = array();
-	public $current_member = - 1;
-	public $members_count = 0;
-	public $in_the_loop = false;
+	public array $members = array();
+	public array $members_ids = array();
+	public int $current_member = - 1;
+	public int $members_count = 0;
+	public bool $in_the_loop = false;
 
-	public $members_latest = array();
+	public array $members_latest = array();
 
 	public function __construct( $args = array(), $parse_request = true ) {
 		$this->_parse_args( $args, $parse_request );
@@ -64,31 +64,31 @@ class Query {
 		);
 
 		if ( $parse_request ) {
-			if ( ! empty( $_GET['orderby'] ) ) {
+			if ( ! empty( $_GET['orderby'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
 				$valid = array_keys( gdmed()->get_sort_orderby_values() );
-				$value = Sanitize::slug( $_GET['orderby'] ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput
+				$value = Sanitize::slug( $_GET['orderby'] ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput,WordPress.Security.NonceVerification.Recommended
 
 				if ( in_array( $value, $valid ) ) {
 					$default['orderby'] = $value;
 				}
 			}
 
-			if ( ! empty( $_GET['order'] ) ) {
+			if ( ! empty( $_GET['order'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
 				$valid = array_keys( gdmed()->get_sort_order_values() );
-				$value = strtoupper( Sanitize::slug( $_GET['order'] ) ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput
+				$value = strtoupper( Sanitize::slug( $_GET['order'] ) ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput,WordPress.Security.NonceVerification.Recommended
 
 				if ( in_array( $value, $valid ) ) {
 					$default['order'] = $value;
 				}
 			}
 
-			if ( ! empty( $_GET['search'] ) ) {
-				$default['search'] = sanitize_title( $_GET['search'] ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput
+			if ( ! empty( $_GET['search'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+				$default['search'] = sanitize_title( $_GET['search'] ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput,WordPress.Security.NonceVerification.Recommended
 			}
 
-			if ( ! empty( $_GET['role'] ) ) {
+			if ( ! empty( $_GET['role'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
 				$valid = array_keys( gdmed()->get_filter_roles_values() );
-				$value = Sanitize::slug( $_GET['role'] ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput
+				$value = Sanitize::slug( $_GET['role'] ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput,WordPress.Security.NonceVerification.Recommended
 
 				if ( in_array( $value, $valid ) ) {
 					$default['role'] = $value;