+++ gd-members-directory-for-bbpress.php
@@ -1,14 +1,14 @@
 <?php
 /**
- * Plugin Name:       GD Members Directory: plugin for WordPress and bbPress
+ * Plugin Name:       membersDirectory for bbPress
  * Plugin URI:        https://plugins.dev4press.com/gd-members-directory-for-bbpress/
  * Description:       Add forum members directory page into bbPress powered forums including members filtering and additional widgets for listing members in the sidebar.
  * Author:            Milan Petrovic
  * Author URI:        https://www.dev4press.com/
  * Text Domain:       gd-members-directory-for-bbpress
- * Version:           2.8
- * Requires at least: 5.9
- * Tested up to:      6.6
+ * Version:           3.0
+ * Requires at least: 6.0
+ * Tested up to:      6.8
  * Requires PHP:      7.4
  * Requires Plugins:  bbpress
  * License:           GPLv3 or later
@@ -31,17 +31,15 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>
  */
 
-use Dev4Press\v51\WordPress;
+use Dev4Press\v54\WordPress;
 
-const GDMED_FILE        = __FILE__;
-const GDMED_PATH        = __DIR__ . '/';
-const GDMED_D4PLIB_PATH = __DIR__ . '/library/';
+define( 'GDMED_FILE', __FILE__ );
+define( 'GDMED_PATH', __DIR__ . '/' );
+define( 'GDMED_URL', plugins_url( '/', __FILE__ ) );
 
-$gdmed_urlname_basic = plugins_url( '/', GDMED_FILE );
+require_once GDMED_PATH . 'vendor/autoload.php';
 
-define( 'GDMED_URL', $gdmed_urlname_basic );
-
-require_once GDMED_D4PLIB_PATH . 'core.php';
+require_once GDMED_PATH . 'vendor/dev4press/library/core.php';
 
 require_once GDMED_PATH . 'core/autoload.php';
 require_once GDMED_PATH . 'core/bridge.php';